# docs/swagger.yaml
definitions:
...
  request.ApiList:
    properties:
      currentPage:
        type: integer
      description:
        type: string
      method:
        type: string
      pageSize:
        type: integer
      path:
        type: string
    type: object
...
  request.CategoryList:
    properties:
      name:
        type: string
      status:
        type: string
    type: object
...
  request.IconList:
    properties:
      name:
        type: string
    type: object
...
  request.ReqCategory:
    properties:
      name:
        type: string
      parent_id:
        type: integer
      sort:
        type: integer
      status:
        type: integer
    required:
    - name
    - sort
    - status
    type: object
...
  request.ReqTag:
    properties:
      name:
        type: string
      sort:
        type: integer
      status:
        type: integer
    required:
    - name
    - sort
    - status
    type: object
  request.TagList:
    properties:
      currentPage:
        type: integer
      name:
        type: string
      pageSize:
        type: integer
      status:
        type: string
    type: object
...
host: 127.0.0.1:9111
info:
  contact: {}
  description: gin框架后台管理
  title: gin框架后台管理demo
  version: "1.0"
paths:
...
  /api/v1/api/getApiList:
    get:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      - description: 菜单名称,状态
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ApiList'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{"list":[],"total":0,"current_page":1,"page_size":20},"msg":"获取API列表成功！"}'
          schema:
            type: string
      summary: API列表
      tags:
      - Api
  /api/v1/api/parent:
    get:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":[],"msg":"获取父级API列表成功！"}'
          schema:
            type: string
      summary: 父级API列表
      tags:
      - Api
...
  /api/v1/category/category:
    post:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分类名称,状态,排序
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ReqCategory'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":null,"msg":"添加文章分类成功！"}'
          schema:
            type: string
      summary: 添加文章分类
      tags:
      - Category
  /api/v1/category/category/:id:
    delete:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":null,"msg":"删除文章分类成功！"}'
          schema:
            type: string
      summary: 删除文章分类
      tags:
      - Category
    put:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分类名称,状态,排序
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ReqCategory'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":null,"msg":"编辑文章分类成功！"}'
          schema:
            type: string
      summary: 编辑文章分类
      tags:
      - Category
  /api/v1/category/getCategoryList:
    get:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分类名称
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CategoryList'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":[],"msg":"获取文章分类列表成功！"}'
          schema:
            type: string
      summary: 文章分类列表
      tags:
      - Category
  /api/v1/category/getSort/:id:
    get:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":[],"msg":"获得子分类sort最大值成功！"}'
          schema:
            type: string
      summary: 子分类sort最大值
      tags:
      - Category
  /api/v1/category/parent:
    get:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":[],"msg":"获取父级文章分类列表成功！"}'
          schema:
            type: string
      summary: 父级文章分类列表
      tags:
      - Category
  /api/v1/icon/getIconList:
    get:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      - description: 名称
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IconList'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":[],"msg":"获取图标列表成功！"}'
          schema:
            type: string
      summary: 图标列表
      tags:
      - Icon
...
  /api/v1/tag/getSelectTagList:
    get:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"获得选择标签列表成功！"}'
          schema:
            type: string
      summary: 选择标签列表
      tags:
      - Tag
  /api/v1/tag/getSort:
    get:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"获得标签sort最大值成功！"}'
          schema:
            type: string
      summary: 标签sort最大值
      tags:
      - Tag
  /api/v1/tag/getTagList:
    get:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      - description: 标签名称,状态
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.TagList'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{"list":[],"total":7,"current_page":1,"page_size":20},"msg":"获取标签列表成功！"}'
          schema:
            type: string
      summary: 标签列表
      tags:
      - Tag
  /api/v1/tag/tag:
    post:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      - description: 标签名称,状态,排序
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ReqTag'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":null,"msg":"添加标签成功！"}'
          schema:
            type: string
      summary: 添加标签
      tags:
      - Tag
  /api/v1/tag/tag/:id:
    delete:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":null,"msg":"删除标签成功！"}'
          schema:
            type: string
      summary: 删除标签
      tags:
      - Tag
    put:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      - description: 标签名称,状态,排序
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ReqTag'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":null,"msg":"编辑标签成功！"}'
          schema:
            type: string
      summary: 编辑标签
      tags:
      - Tag
...
swagger: "2.0"